requirements.txt
CouchDB==0.9
Flask==0.10.1
Flask-Login==0.2.10
Jinja2==2.7.2
MarkupSafe==0.19
Werkzeug==0.9.4
backports.ssl-match-hostname==3.4.0.2
evernote==1.25.0
google-api-python-client==1.2
httplib2==0.8
itsdangerous==0.23
oauth2==1.5.211
python-dateutil==2.2
pytz==2014.1
redis==2.7.1
redis-simple-cache==0.0.6
six==1.6.1
tornado==3.2
wsgiref==0.1.2
